Enum jsonrpc_server_utils::reactor::UninitializedRemote

pub enum UninitializedRemote {
    Shared(Remote),
    Unspawned,
}

Possibly uninitialized event loop remote.

Variants
Shared(Remote)

Shared instance of remote.

Unspawned

Event Loop should be spawned by the transport.

Methods

impl UninitializedRemote

pub fn initialize(self) -> Result<Remote>

Initializes remote. In case there is no shared remote, will spawn a new event loop. Dropping Remote closes the loop.

pub fn init_with_name<T: Into<String>>(self, name: T) -> Result<Remote>

Initializes remote. In case there is no shared remote, will spawn a new event loop. Dropping Remote closes the loop.