Struct jsonrpc_server_utils::reactor::RpcEventLoop

pub struct RpcEventLoop { /* fields omitted */ }

A handle to running event loop. Dropping the handle will cause event loop to finish.

Methods

impl RpcEventLoop

fn spawn() -> Result<Self>

Spawns a new thread with the EventLoop.

fn with_name(name: Option<String>) -> Result<Self>

Spawns a new named thread with the EventLoop.

fn remote(&self) -> Remote

Get remote for this event loop.

fn wait(self) -> Result<()>

Blocks current thread and waits until the event loop is finished.

fn close(self)

Finishes this event loop.