Struct jsonrpc_server_utils::SuspendableStream

pub struct SuspendableStream<S> { /* fields omitted */ }

Incoming is a stream of incoming sockets Polling the stream may return a temporary io::Error (for instance if we can't open the connection because of "too many open files" limit) we use for_each combinator which:

Runs for every Ok(socket)

Stops on the FIRST Err() So any temporary io::Error will cause the entire server to terminate. This wrapper type for tokio::Incoming stops accepting new connections for a specified amount of time once an io::Error is encountered

Methods

impl<S> SuspendableStream<S>

pub fn new(stream: S) -> Self

construct a new Suspendable stream, given tokio::Incoming and the amount of time to pause for.